requirements.txt
colorama==0.3.9
pathfinding==0.0.1
flask_socketio